Ignore nonexistent builds from Travis widget
Why:

We're only interested in the status of repos that actually have builds running on Travis.

How:
I refactored and simplified the Travis job to only use the repo's `last_build_status` as opposed to looking up branches. I'm not sure why the original code was digging into branches, but what I did does the job.

Basically, I'm getting an array of all of 18F's repos that are listed on Travis, except the ones that don't have a `last_build_id`. That means that no builds are running on Travis for that repo, and we can ignore it.

In some cases, some repos were running on the public Travis, but have sinced moved to the private Travis. Since those builds are old and not relevant anymore, we want to exclude them as well, but this is a manual process via the `TRAVIS_IGNORED_REPOS` env var.

In addition, I added a new widget that lists all testable repos that should have builds running on Travis, but do not. Testable repos are obtained via the `team-api`.