Add widget to show repos missing .about.yml
To track adoption of the `.about.yml` standard, we thought it would be helpful to add a widget that shows which repos still have not added it.

In order to make it easy to do this with the inherited code that we forked, I made the `ORGS` env var singular, since we're only using this project for one GitHub organization.

I also moved the GitHub stats widgets to their own page at /github_stats to reduce clutter on the /default endpoint.

Closes #15.