Add CircleCI build status widget
The Circle CI API does not currently provide a way to get a list of just
the active repos, so we have to use the GitHub API to get a list of all
repos that belong to an organization, then make a separate Circle API
call for each one (over 300 for 18F), just to determine whether
Circle CI has builds for that repo.

In order to minimize the amount of times we have to do this, I created
a new job that runs once a day and stores this list in an env var,
which can then be reused by the Circle CI widget, which runs every 5
minutes.

I also did the same optimization for the list of GitHub org repos.

To support this new optimization, I adjusted the start time of each
job so that it would allow enough time to pass for the env vars to get
populated.

Closes #14.